+++ crates/ibc/src/applications/transfer/events.rs
@@ -1,6 +1,6 @@
 use crate::applications::transfer::acknowledgement::TokenTransferAcknowledgement;
 use crate::applications::transfer::{Amount, PrefixedDenom, MODULE_ID_STR};
-use crate::events::ModuleEvent;
+use crate::core::events::ModuleEvent;
 use crate::prelude::*;
 use crate::signer::Signer;
