+++ crates/ibc/src/core/context.rs
@@ -33,6 +33,7 @@ use core::time::Duration;
 
 use ibc_proto::google::protobuf::Any;
 
+use crate::core::events::IbcEvent;
 use crate::core::ics02_client::client_state::ClientState;
 use crate::core::ics02_client::consensus_state::ConsensusState;
 use crate::core::ics03_connection::connection::ConnectionEnd;
@@ -64,7 +65,6 @@ use crate::core::{
     ics24_host::identifier::ClientId,
     ics26_routing::{error::RouterError, msgs::MsgEnvelope},
 };
-use crate::events::IbcEvent;
 use crate::timestamp::Timestamp;
 use crate::Height;
