+++ crates/ibc/src/core/context/chan_close_confirm.rs
@@ -7,7 +7,7 @@ use crate::core::ics04_channel::handler::chan_close_confirm;
 use crate::core::ics04_channel::msgs::chan_close_confirm::MsgChannelCloseConfirm;
 use crate::core::ics26_routing::context::ModuleId;
 
-use crate::events::{IbcEvent, MessageEvent};
+use crate::core::events::{IbcEvent, MessageEvent};
 
 use super::{ContextError, ExecutionContext, ValidationContext};
 
@@ -100,11 +100,11 @@ mod tests {
 
     use crate::applications::transfer::MODULE_ID_STR;
     use crate::core::context::chan_close_confirm::chan_close_confirm_execute;
+    use crate::core::events::{IbcEvent, MessageEvent};
     use crate::core::ics04_channel::msgs::chan_close_confirm::test_util::get_dummy_raw_msg_chan_close_confirm;
     use crate::core::ics04_channel::msgs::chan_close_confirm::MsgChannelCloseConfirm;
     use crate::core::ics26_routing::context::ModuleId;
     use crate::core::ValidationContext;
-    use crate::events::{IbcEvent, MessageEvent};
 
     use crate::core::ics03_connection::connection::ConnectionEnd;
     use crate::core::ics03_connection::connection::Counterparty as ConnectionCounterparty;
