+++ crates/ibc/src/core/ics02_client/handler/create_client.rs
@@ -1,6 +1,6 @@
 //! Protocol logic specific to processing ICS2 messages of type `MsgCreateClient`.
 
-use crate::events::MessageEvent;
+use crate::core::events::MessageEvent;
 use crate::prelude::*;
 
 use crate::core::context::ContextError;
@@ -13,11 +13,11 @@ use crate::core::ExecutionContext;
 
 use crate::core::ValidationContext;
 
+use crate::core::events::IbcEvent;
 use crate::core::ics02_client::error::ClientError;
 use crate::core::ics02_client::events::CreateClient;
 use crate::core::ics02_client::msgs::create_client::MsgCreateClient;
 use crate::core::ics24_host::identifier::ClientId;
-use crate::events::IbcEvent;
 
 pub(crate) fn validate<Ctx>(ctx: &Ctx, msg: MsgCreateClient) -> Result<(), ContextError>
 where
