+++ crates/ibc/src/core/ics04_channel/handler/send_packet.rs
@@ -1,3 +1,5 @@
+use crate::core::events::IbcEvent;
+use crate::core::events::MessageEvent;
 use crate::core::ics04_channel::channel::Counterparty;
 use crate::core::ics04_channel::channel::State;
 use crate::core::ics04_channel::commitment::compute_packet_commitment;
@@ -11,8 +13,6 @@ use crate::core::ics24_host::path::ClientConsensusStatePath;
 use crate::core::ics24_host::path::CommitmentPath;
 use crate::core::ics24_host::path::SeqSendPath;
 use crate::core::ContextError;
-use crate::events::IbcEvent;
-use crate::events::MessageEvent;
 use crate::prelude::*;
 use crate::timestamp::Expiry;
 
@@ -142,6 +142,7 @@ mod tests {
 
     use test_log::test;
 
+    use crate::core::events::IbcEvent;
     use crate::core::ics02_client::height::Height;
     use crate::core::ics03_connection::connection::ConnectionEnd;
     use crate::core::ics03_connection::connection::Counterparty as ConnectionCounterparty;
@@ -153,7 +154,6 @@ mod tests {
     use crate::core::ics04_channel::packet::Packet;
     use crate::core::ics04_channel::Version;
     use crate::core::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
-    use crate::events::IbcEvent;
     use crate::mock::context::MockContext;
     use crate::timestamp::Timestamp;
     use crate::timestamp::ZERO_DURATION;
