+++ crates/ibc/src/core/mod.rs
@@ -31,7 +31,7 @@ pub mod ics24_host;
 pub mod ics26_routing;
 
 pub mod context;
-
+pub mod events;
 pub mod handler;
 
 pub use handler::dispatch;
